requirements.txt
asgiref==3.3.1
Django==3.1.5
image==1.5.33
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
